fix mdns crash on scoped IPv6 link-local
Before this change the library will always crash with "SocketException: no route to host" when used on IPv6 interface with a scoped address e.g.[FE80::1%1].

The root cause was because the multicast address for IPv6 didn't specify a scope ID for such cases.

This commit fixes it by first checking if the local endpoint is an IPv6 link local, if true, it will create and use a scoped multicast address for mdns. The resulting multicast address is then stored for reuse.